# serialize: Fixed handling of nested object references