PHP 8.1+ does not serialize object references correctly under certain circumstances.

See tests/serialize/serialization_objects_019.phpt

The bug was originally introduced in commit 6c5942f, and the problematic line was last modified in commit bb0b4eb9. (Fixes oss-fuzz #44954)

The testcase from bb0b4eb9 still passes.

I am not super familiar with the php codebase, but the removed IF looks suspicious and fixes my issue. I initially planned to just open an issue, but then I wrote the testcase and figured I might as well just submit the (likely) fix.

Someone else should obviously triple-check this. Especially since the original change was due to optimizations.