@iluuu1994 2023-05-29
@dstogov That sounds reasonable. It would need to account for recursive arrays, as you could add one more level of nesting to make the RC go back to 1:

```mermaid
graph TD;
    $r-->a;
    $r-->b;
    a-->array1_r2;
    b-->array1_r2;
    array1_r2-->array2_rc1;
    array2_rc1-->$s;
```

If you like I can try that to see if it makes any difference.