@dstogov 2023-05-29
@iluuu1994 it may make sense to try this. That optimization might really make improvement. On the other hand, passing more arguments and additional checks may make degradation for the more  common cases.